requirements.txt
pandas
XlsxWriter
pyodbc
azure-storage-file-share
python_dotenv
requests
openpyxl